♻️ 코드 수정: Lecture, LectureByStudent 순환 참조 Facade 이용해서 해결
