+++ LearnsMate/src/main/java/intbyte4/learnsmate/lecture/controller/LectureController.java
@@ -1,5 +1,6 @@
 package intbyte4.learnsmate.lecture.controller;
 
+import intbyte4.learnsmate.facade.LectureFacade;
 import intbyte4.learnsmate.lecture.domain.dto.LectureDTO;
 import intbyte4.learnsmate.lecture.domain.vo.request.RequestEditLectureInfoVO;
 import intbyte4.learnsmate.lecture.domain.vo.request.RequestRegisterLectureVO;
@@ -27,6 +28,7 @@ public class LectureController {
 
     private final LectureService lectureService;
     private final LectureMapper lectureMapper;
+    private final LectureFacade lectureFacade;
 
     @Operation(summary = "강의 정보 전체 조회")
     @GetMapping
@@ -66,7 +68,7 @@ public class LectureController {
     @Operation(summary = "강의 삭제")
     @PatchMapping("/{lectureCode}/status")
     public ResponseEntity<ResponseRemoveLectureVO> removeLecture(@PathVariable("lectureCode")  Long lectureCode) {
-        LectureDTO removedLecture = lectureService.removeLecture(lectureCode);
+        LectureDTO removedLecture = lectureFacade.removeLecture(lectureCode);
         return ResponseEntity.status(HttpStatus.OK).body(lectureMapper.fromDtoToRemoveResponseVO(removedLecture));
     }
