+++ LearnsMate/src/main/java/intbyte4/learnsmate/lecture/service/LectureService.java
@@ -2,6 +2,7 @@ package intbyte4.learnsmate.lecture.service;
 
 import intbyte4.learnsmate.issue_coupon.domain.dto.IssueCouponDTO;
 import intbyte4.learnsmate.lecture.domain.dto.LectureDTO;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -11,5 +12,7 @@ public interface LectureService {
     List<LectureDTO> getLecturesByTutorCode(Long tutorCode);
     LectureDTO registerLecture(LectureDTO lectureDTO, List<Integer> lectureCategoryCodeList);
     LectureDTO updateLecture(Long lectureId, LectureDTO requestEditLectureInfoVO);
-    LectureDTO removeLecture(Long lectureCode);
+
+//    @Transactional
+//    LectureDTO removeLecture(Long lectureCode);
 }
