+++ LearnsMate/src/main/java/intbyte4/learnsmate/lecture_by_student/repository/LectureByStudentRepository.java
@@ -19,7 +19,7 @@ public interface LectureByStudentRepository extends JpaRepository<LectureByStude
 
     @Query("SELECT l.lectureByStudentCode FROM lecture_by_student l WHERE l.lecture.lectureCode = :lectureCode")
     List<Long> findLectureByStudentCodesByLectureCode(@Param("lectureCode") Long lectureCode);
-    long countByLectureAndRefundStatus(Lecture lecture, boolean refundStatus);
+//    long countByLectureAndOwnStatus(Lecture lecture, boolean refundStatus);
 
     LectureByStudent findByLecture(Lecture lecture);
